Implemented Course Section create and delete interface.
NEW
* Implemented the `create` and `delete` functionality for the
  Course Section administrative interface.
* Created the CourseSection controller.

CHANGES
* Created a new route filter to deal with setting an appropriate
  school session and keeping track of it.
* Updated the database seed so that there is a default school
  session.
* Changed the School Session controller so that the default
  session cannot be deleted.

+++ app/controllers/AdminSessionController.php
@@ -76,6 +76,13 @@ class AdminSessionController extends BaseController {
       return Redirect::action('AdminSessionController@index');
     }
 
+    if( $session->id == 1 ) {
+      Session::flash('action_success', false);
+      Session::flash('action_message', 'You cannot remove the default school session.');
+
+      return Redirect::action('AdminSessionController@index');
+    }
+
     $sessionCode = $session->code;
 
     $session->delete();
